Module geckoservo::glue

Functions

Servo_AnimationValue_AddRef

Servo_AnimationValue_GetOpacity

Servo_AnimationValue_GetTransform

Servo_AnimationValue_Release

Servo_AnimationValue_Serialize

Servo_AnimationValues_Interpolate

Servo_AnimationValues_Populate
Takes a ServoAnimationValues and populates it with the animation values corresponding to a given property declaration block

Servo_AnimationValues_Uncompute

Servo_AssertTreeIsClean

Servo_CSSSupports

Servo_CSSSupports2

Servo_ComputedValues_AddRef

Servo_ComputedValues_GetForAnonymousBox

Servo_ComputedValues_Inherit

Servo_ComputedValues_Release

Servo_CssRules_AddRef

Servo_CssRules_DeleteRule

Servo_CssRules_GetStyleRuleAt

Servo_CssRules_InsertRule

Servo_CssRules_ListTypes

Servo_CssRules_Release

Servo_DeclarationBlock_AddPresValue

Servo_DeclarationBlock_AddRef

Servo_DeclarationBlock_Clone

Servo_DeclarationBlock_Count

Servo_DeclarationBlock_CreateEmpty

Servo_DeclarationBlock_Equals

Servo_DeclarationBlock_GetCssText

Servo_DeclarationBlock_GetNthProperty

Servo_DeclarationBlock_GetPropertyIsImportant

Servo_DeclarationBlock_GetPropertyValue

Servo_DeclarationBlock_GetPropertyValueById

Servo_DeclarationBlock_Release

Servo_DeclarationBlock_RemoveProperty

Servo_DeclarationBlock_RemovePropertyById

Servo_DeclarationBlock_SerializeOneValue

Servo_DeclarationBlock_SetProperty

Servo_DeclarationBlock_SetPropertyById

Servo_Element_ClearData

Servo_Element_GetSnapshot

Servo_ImportRule_AddRef

Servo_ImportRule_GetSheet

Servo_ImportRule_Release

Servo_Initialize

Servo_NoteExplicitHints

Servo_ParseProperty

Servo_ParseStyleAttribute

Servo_ResolvePseudoStyle

Servo_ResolveStyle

Servo_ResolveStyleLazily

Servo_RestyleWithAddedDeclaration

Servo_Shutdown

Servo_StyleRule_AddRef

Servo_StyleRule_Debug

Servo_StyleRule_GetCssText

Servo_StyleRule_GetSelectorText

Servo_StyleRule_GetStyle

Servo_StyleRule_Release

Servo_StyleRule_SetStyle

Servo_StyleSet_AppendStyleSheet

Servo_StyleSet_Drop

Servo_StyleSet_FillKeyframesForName

Servo_StyleSet_FlushStyleSheets

Servo_StyleSet_Init
See the comment in Device to see why it's ok to pass an owned reference to the pres context (hint: the context outlives the StyleSet, that holds the device alive).

Servo_StyleSet_InsertStyleSheetBefore

Servo_StyleSet_NoteStyleSheetsChanged

Servo_StyleSet_PrependStyleSheet

Servo_StyleSet_RebuildData

Servo_StyleSet_RemoveStyleSheet

Servo_StyleSheet_AddRef

Servo_StyleSheet_ClearAndUpdate

Servo_StyleSheet_Empty

Servo_StyleSheet_FromUTF8Bytes

Servo_StyleSheet_GetRules

Servo_StyleSheet_HasRules

Servo_StyleSheet_Release

Servo_StyleWorkerThreadCount

Servo_TakeChangeHint

Servo_TraverseSubtree
Traverses the subtree rooted at root for restyling. Returns whether a Gecko post-traversal (to perform lazy frame construction, or consume any RestyleData, or drop any ElementData) is required.